Reworked to Allow On-Demand Load of Any Directory
Essentially, we've moved away from the old approach where the loader was rigged up to fire once specifically after the Fighter.pac injects finished on a specific target directory. It's now a reusable loader which can now instead be invoked *whenever*, on *whatever* directory you want! To invoke, simply write the address of the target directory path string to the source path address, and that'll kickstart the rest of the loader, running it on that folder!

Also moved to a slightly different main hookpoint, just to get off the one used by the Syriinge template project, since I imagine that could easily lead to conflicts lol.