Code Buffer is Now Dynamically Allocated
Rather than constantly hoarding 8kb of space in the Syriinge heap for no reason, we now have a system in place for dynamically allocating our buffers at application time!

Currently things are set up such that the Fighter Inject application explicitly asks for a 8kb allocation in the OverlayFighter1 heap, though that's easily changeable if that turns out not to work actually.