Avoided Unneeded (Un)RegisterCallback Templating
Previously the registerCallback and unregisterCallback functions were just template functions specialized off the type of callback being registered. This meant that every time we added a new callback type, a new copy of the entire function was being created. This was unnecessary though, because all we were doing really is adding a pointer to a specific vector of pointers; we don't need to care what type of pointer they are actually.

So, the template functions are now just inlined functions which cast the incoming vector and callback pointers to Vector<void*>* and void* pointers respectively, and passes those along to the actual, singular register and unregister functions! Saves a good deal of space in the .rel!