Began Working on Reflect On Hit Functionality
This is currently in a pretty incomplete state; all it does currently is set player hurtboxes to enable reflection whenever a hitbox is out, which is *kinda* like the goal and may be part of the final solution but is definitely incomplete. In particular characters with large disjoints like swordies get very little use out of it, and ultimately most attacks will clank before they reach the hurtbox to be deflected in the first place, so just overall this'll need reworking.