RastVar Stack Adjustment Installs on SoraMelee Load
Had this set to proc on selchar module load before just cuz I wasn't sure if the ModuleLoadEvent subscription system fired before or after the module had completed loading. Can now confirm it fires afterwards, so it's safe to just fire on SORA_MELEE load (which in testing seems to functionally be immediately cuz it's loaded by the time Syriinge pulls this module in). Also made the write addresses relative to the REL's location in the same way the actual main hook is. Not sure if that'll ever matter in practice, but not much reason not to do it I don't think lol.