Optimizations to FighterHook Callback Vectors
Converted all the callback vectors to Vector<void*> so we don't have to waste space in the REL on duplicate templated member functions for every callback type.

Also set up a basic #define to disable the ftOutsideEventObserver, since we're not actually using it at the moment and it nearly doubles the size of the REL for no real benefit.