Moved Meter Tracking to Internal Arrays
The old approach of storing these as PSA Variables was cool in concept but just isn't really safe given the unpredictable usage of PSA Vars in any given character's PSA. Technically even the RA-Bits I'm using aren't strictly *safe*, but the odds of those conflicting are a lot lower just because of how many they are and how infrequently they seem to be used in PSAs.
Also learned you can get a Fighter's Entry ID from the struct itself, so simplified the getFighterSlotNo function using that.