Fixed Incorrect GetFighterSlotNo Use
The previous way of doing this I had was getting the bottom byte of the entryID; but apparently that isn't actually *Player* ID, it's "Slot" ID, which is to say the slot they occupy in the current match. Made a separate GetFighterPlayerNo function to do what I'd intended, switched to using that where relevant.

Also, got rid of the unnecessary re-definitions of the gameGlobal and ftManager globals, using the BrawlHeaders ones instead now!