Improved Slime Cancel SFX, GFX, Perfect Cancel Adv
Learned you can actually set the scale of GFX in each axis, so switched to using the Raygun Shot effect with Z-Scale set to 0.1 which looks a lot better imo. Also lets us move the old main effect to the stock-gain GFX slot, which helps feedback on that end a lot.

Additionally, performing a Slime Cancel during hitstop now gives you frame advantage equal to that hitstop instance's original length, before being extended by the cancel. This loosens up the timing necessary to get the full benefit and makes the total frame advantage gained on any given hit consistent, which should make it easier to learn and rely on the resulting timing.