Initial CM_Addon Interface System Impl.
Establishes a new system for loading in the auto-generated "Index Files" that the Code Menu Builder can spit out for Addons. These contain the INDEX values of every included line within the Addon, which when coupled with careful control over the order in which Lines are declared *within* an Addon can allow proper communication with Syriinge.