Misc. Rocket Jumps Polish
Put some restrictions on when you're allowed to Rocket Jump, and reworked the logic for killing the hitbox to be a bit more intuitive (should function the same). Also pushed effects proc one frame back to fix an oversight where the jump effects could proc before the hitbox was actually active.