Impl. Hook API Wrapper for Pre-0.6.0 Back-Compat
Implements a wrapper over the Hooking API which allows us to easily switch between the old system of linking against sy_core.map, and the new system using the CoreApi object. In practice this should allow the plugins in this repo to remain compatible with pre-0.6.0 Syriinge, which means builds for one reason or another locked to that older version may continue to use them!