Split Wavedash Mechanics out of Rivals Mode
This is the first of many rounds of this, to separate each component of the Rivals Mode blob into its own toggleable setting. This *is* going to ultimately make the plugin a fair bit bigger though, so I think the time to investigate loading and unloading plugins is sooner rather than later.

Also wanna see about optimizing away the duplicate code menu lines for the passive mechanics especially, would be nice.