Added MeleeOnUpdateCallback
Note: Currently doesn't fire in SSE, though none of the others do either, so not sure how big of a deal it actually is? Would ideally like to fix this one though; match start and stop stuff you wouldn't *expect* to work in SSE necessarily, this one you probably would.