Mechanics May Now be Enabled or Disabled Mid-Game
Implemented an additional set of flags that indicate whether or not a change to whether or not a mechanic is enabled has occurred between the previous frame and the current one. This in combination with the enabled/disabled masks allow you to detect when a mechanic turns on or off from within the Fighter Update callback, which we use now to have Active Mechanics reset their meter upon activation mid-match!