Reimp'd Once-Per-Airtime Walljump Mechanic
Upon thinking further about this, I realized I'd actually interpreted the mechanic incorrectly: it isn't so much that you can only walljump once per airtime, it's that you're required to walljump in *opposing directions* (which just happens to be near-impossible in one airtime on most legal stages)! With that in mind, I've rewritten most of how that's handled, and in doing so allowed characters to walljump back and forth between opposing walls again like they can in Rivals. Code Menu option and internal labelling changed to reflect this as well.