Fixed Timing Conflict for FrameHeap 01 Banks
Triggering the load for the common banks on exiting scMelee while on a stage with an active soundbank was causing banks to load into FH01 just before the stage system cleared that Frame Heap down to nothing, effectively undoing the load.

Moved instead to a system oriented around proper scene changes, which allows us to wait till after scMemoryChange to do our loads.