Removed Loading and Unloading of Narration_Melee
This was a remnant from when the intention was swapping Menu and Narration_Melee in and out in the Common Heap, but that's no longer necessary because Menu is loading in FH01 now.
While they were both loading in 0x00 it helped in that they were never loaded at the same time, so we only had to budget for the bigger of the two, but now that it's just the one it doesn't actually do anything for us.